modified:   src/goroutine.go 	modified:   test/Makefile 	new file:   test/goroutine.h 	modified:   test/goroutine.hh 	new file:   test/pingpong.c

+++ src/goroutine.go
@@ -56,7 +56,7 @@ func _goroutine_yield() {
 var chanPool = map[uintptr](*chan uintptr){}
 
 //export _goroutine_new_channel
-func _goroutine_new_channel(count int) uintptr {
+func _goroutine_new_channel(count uintptr) uintptr {
 	ch := make(chan uintptr, count)
 	ptr := uintptr(unsafe.Pointer(&ch))
 	chanPool[ptr] = &ch
